+++ src/components/Input/Checkbox.tsx
@@ -1,13 +1,28 @@
 import { styled } from "styled-components";
+import { TodoManager } from "../../util/TodoManager";
+import { LocalStorageManager } from "../../util/localStorageManager";
+import { TodoTypes } from "../../types/TodoTypes";
+import { Dispatch } from "react";
 
-const Checkbox = ({
-  completed,
-  onChange,
-}: {
+interface CheckBoxTypes {
+  todoId: number;
+  setInitTodo: Dispatch<React.SetStateAction<TodoTypes[]>>;
   completed: boolean;
-  onChange: () => void;
-}) => {
-  return <CheckboxComponent onChange={onChange} checked={completed} />;
+}
+const Checkbox = ({ todoId, setInitTodo, completed }: CheckBoxTypes) => {
+  const todoManager = new TodoManager();
+  const localStorageManager = new LocalStorageManager();
+
+  const handleChange = () => {
+    todoManager.complete(todoId, localStorageManager);
+    setInitTodo((prevTodos) =>
+      prevTodos.map((todo) =>
+        todo.id === todoId ? { ...todo, completed: !todo.completed } : todo
+      )
+    );
+  };
+
+  return <CheckboxComponent onChange={handleChange} checked={completed} />;
 };
 
 export default Checkbox;
